requirements.txt
############################################################
# STREAMLIT (Dashboard)
############################################################
streamlit==1.39.0

############################################################
# MODELAGEM / LLMs
############################################################
openai>=1.40.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.0

############################################################
# CREWAI (Agentes)
############################################################
crewai>=0.51.0
crewai-tools>=0.4.0

############################################################
# MANIPULAÇÃO DE DADOS
############################################################
pandas>=2.2.0
numpy>=1.26.0
tabulate>=0.9.0

############################################################
# VISUALIZAÇÃO (Dashboard)
############################################################
plotly>=5.24.0

############################################################
# SCRAPING
############################################################
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=5.2.0

############################################################
# UTILITÁRIOS
############################################################
# Evita problemas com versões do pip / wheel / setuptools
setuptools>=69.0.0
wheel>=0.43.0
